+++ src/nespresso/recsys/searchbase/search.py
@@ -1,10 +1,14 @@
+import logging
 from dataclasses import dataclass
 from typing import Any
+
+from aiogram import types
 
 from nespresso.recsys.searchbase.document import DocAttribute
 from nespresso.recsys.searchbase.index import INDEX_NAME, client
 
-SCROLL_TIMEOUT = "30m"  # alive for 30 minutes
+_TIMEOUT = "30m"  # alive for 30 minutes
+_ITEMS = 5
 
 
 @dataclass
@@ -13,7 +17,7 @@ class SearchItem:
     score: float
 
     @classmethod
-    def FromDict(cls, hit: dict[Any, Any]) -> "SearchItem":
+    def FromHit(cls, hit: dict[Any, Any]) -> "SearchItem":
         return cls(
             nes_id=int(hit["_id"]),
             score=float(hit["_score"]),
@@ -21,59 +25,141 @@ class SearchItem:
 
 
 @dataclass
-class SearchResult:
-    items: list[SearchItem] | None = None
-    scroll_id: str | None = None
-    expired: bool = False
+class SearchPage:
+    number: int
+    items: list[SearchItem]
+    scroll_id: str
+
+    @staticmethod
+    def _ExtractItems(response: dict[Any, Any]) -> list[SearchItem]:
+        return [SearchItem.FromHit(hit) for hit in response["hits"]["hits"]]
 
     @classmethod
-    def FromDict(cls, response: dict[Any, Any]) -> "SearchResult":
-        hits = response["hits"]["hits"]
-
-        items = [SearchItem.FromDict(hit) for hit in hits]
-        scroll_id = response["_scroll_id"]
-
-        return cls(items=items, scroll_id=scroll_id)
-
-
-async def HybridSearch(text: str) -> SearchResult:
-    attr = DocAttribute.FromText(text)
-
-    search_body = {
-        "size": 5,
-        "_source": False,
-        "query": {
-            "bool": {  # composite query
-                "should": [  # scores are summed
-                    {"match": {"mynes_keywords": attr.keywords}},
-                    {"knn": {"mynes_embedding": {"vector": attr.embedding, "k": 5}}},
-                    {"match": {"cv_keywords": attr.keywords}},
-                    {"knn": {"cv_embedding": {"vector": attr.embedding, "k": 5}}},
-                ]
-            }
-        },
-    }
-
-    response = await client.search(
-        index=INDEX_NAME,
-        body=search_body,
-        scroll=SCROLL_TIMEOUT,
-    )
-
-    return SearchResult.FromDict(response)
-
-
-async def ScrollThroughResults(scroll_id: str) -> SearchResult:
-    try:
-        response = await client.scroll(
-            scroll_id=scroll_id,
-            scroll=SCROLL_TIMEOUT,  # refresh
+    def FromResponse(cls, response: dict[Any, Any], number: int) -> "SearchPage":
+        return cls(
+            number=number,
+            items=SearchPage._ExtractItems(response),
+            scroll_id=response["_scroll_id"],
         )
-    except Exception:
-        return SearchResult(expired=True)
-
-    return SearchResult.FromDict(response)
 
 
-async def FinishScrolling(scroll_id: str) -> None:
-    await client.clear_scroll(scroll_id=scroll_id)
+class ScrollingSearch:
+    def __init__(self) -> None:
+        self.pages: list[SearchPage] = []
+        self.index: int = 0
+        self.expired = False
+
+    def _CreateBody(self, message: types.Message) -> dict[Any, Any]:
+        if not message.text:
+            raise ValueError("Expected message.text to be non-empty")
+
+        attr = DocAttribute.FromText(message.text)
+
+        logging.info(f"chat_id={message.chat.id} :: Query keywords: '{attr.keywords}'")
+
+        body = {
+            "size": _ITEMS,
+            "_source": False,
+            "query": {
+                "bool": {  # composite query
+                    "should": [  # scores are summed
+                        {"match": {"mynes_keywords": attr.keywords}},
+                        {
+                            "knn": {
+                                "mynes_embedding": {
+                                    "vector": attr.embedding,
+                                    "k": _ITEMS,
+                                }
+                            }
+                        },
+                        {"match": {"cv_keywords": attr.keywords}},
+                        {
+                            "knn": {
+                                "cv_embedding": {
+                                    "vector": attr.embedding,
+                                    "k": _ITEMS,
+                                }
+                            }
+                        },
+                    ]
+                }
+            },
+        }
+
+        return body
+
+    def _CurrentPage(self) -> SearchPage:
+        return self.pages[self.index]
+
+    async def HybridSearch(self, message: types.Message) -> SearchPage | None:
+        if self.pages:
+            raise ValueError("HybridSearch() was called more than once.")
+
+        body = self._CreateBody(message)
+
+        response = await client.search(
+            index=INDEX_NAME,
+            body=body,
+            scroll=_TIMEOUT,
+        )
+
+        page = SearchPage.FromResponse(
+            response=response,
+            number=self.index,
+        )
+        self.pages = [page]
+
+        # TODO: check for score (if it is high enough) and output `None`
+
+        return self._CurrentPage()
+
+    async def ScrollBackward(self) -> SearchPage | None:
+        if self.index == 0:
+            return None
+
+        self.index -= 1
+
+        return self._CurrentPage()
+
+    async def ScrollForward(self) -> SearchPage | None:
+        if not self.pages:
+            raise ValueError("HybridSearch() must be called before scrolling forward.")
+
+        if self.index < self.pages[-1].number:
+            self.index += 1
+            return self._CurrentPage()
+
+        if self.expired:
+            return None
+
+        try:
+            response = await client.scroll(
+                scroll_id=self.pages[-1].scroll_id,
+                scroll=_TIMEOUT,  # refresh
+            )
+        except Exception:
+            self.expired = True
+            return None
+
+        page = SearchPage.FromResponse(
+            response=response,
+            number=self.index + 1,
+        )
+
+        if not page.items:
+            return None
+
+        # TODO: check for score (if it is high enough) and output `None`
+
+        self.pages.append(page)
+        self.index += 1
+
+        return self._CurrentPage()
+
+    async def FinishScrolling(self) -> None:
+        if not self.pages:
+            raise ValueError(
+                "HybridSearch() must be called before finishing scrolling."
+            )
+
+        await client.clear_scroll(scroll_id=self.pages[-1].scroll_id)