+++ CHANGELOG.md
@@ -23,6 +23,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.1.0] - 2025-06-04 - Vlad
 ### Added
+- Scrolling search framework
+
+## [0.1.0] - 2025-06-04 - Vlad
+### Added
 - Email verification of NES
 - Extracting of keywords from text